// src/main/components/safe-box/usecases/add-users/add-users-request-dto.ts
export interface IAddUsersRequestDTO {
  usuarios_leitura: string[];
  usuarios_escrita: string[];
  tipo: string;
  usuarios_leitura_deletado: string[];
  usuarios_escrita_deletado: string[];
  criptografia: string;
  nome: string;
  anexos: string[];
  descricao: string;
  conteudo: string[];
  organizacao: string;
  _id: string;
}